+++ Source/OxyPlot.Wpf.Shared/Tracker/TrackerControl.cs
@@ -41,6 +41,12 @@ namespace OxyPlot.Wpf
                 new PropertyMetadata(Visibility.Visible));
 
         /// <summary>
+        /// Identifies the <see cref="LineThickness"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty LineThicknessProperty = DependencyProperty.Register(
+            nameof(LineThickness), typeof(double), typeof(TrackerControl), new PropertyMetadata(1.0));
+
+        /// <summary>
         /// Identifies the <see cref="LineStroke"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty LineStrokeProperty = DependencyProperty.Register(
@@ -189,6 +195,15 @@ namespace OxyPlot.Wpf
         {
             get => (Visibility)this.GetValue(VerticalLineVisibilityProperty);
             set => this.SetValue(VerticalLineVisibilityProperty, value);
+        }
+
+        /// <summary>
+        /// Gets or sets LineThickness.
+        /// </summary>
+        public double LineThickness
+        {
+            get => (double)this.GetValue(LineThicknessProperty);
+            set => this.SetValue(LineThicknessProperty, value);
         }
 
         /// <summary>