+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+- Add LineThickness property to TrackerControl (#1831)
+
+### Changed
+
+### Removed
+
 ### Fixed
 - WPF - OxyPlot doesn't render inside a Popup (#1796)
 - Odd behavior of zooming of Logarithmic axis in Cartesian plot (#1825)
+- SkiaSharp.WPF - OxyPlot doesn't render inside an ElementHost (#1800)
+- NullReference in SkiaSharp WPF renderer if UIElement has no PresentationSource  (#1798)
 
 ## [2.1.0] - 2021-10-02